sync builder is moved to api package copyright fixed for sdk
